Fixed bug where new checkTarget function was not available for others than walking entities
